updated for Xcode 10-3

+++ Packages/Node-1.0.1/Sources/Node/Core/Node+Init.swift
@@ -8,7 +8,7 @@ extension Node {
     }
 
     public init(_ int: Int) {
-        self = .number(Number(int))
+        self = .number(Number(UInt(int)))
     }
 
     public init(_ double: Double) {
